driver: Replace pointer casting with map_physmem()
Replace pointer casting with map_physmem() and give the permission as the
pointer may not be an unsigned long integer and may have intrinsic
permission on architecture like CHERI.

+++ include/ns16550.h
@@ -68,7 +68,7 @@ enum ns16550_flags {
  * @bdf:		PCI slot/function (pci_dev_t)
  */
 struct ns16550_plat {
-	ulong base;
+	uintptr_t base;
 	ulong size;
 	int reg_width;
 	int reg_shift;
